Revert "Revert "Added UV and texture support into the engine""
This reverts commit 308ae4a5f1eed2ef3555f0d72cc56507a0d62c71.

+++ Source/Object/Mesh/2D/NRE_Mesh2D.cpp
@@ -33,7 +33,7 @@
 
                 add(0, vData, 18);
 
-                GLshort uvData[] = {
+                GLfloat uvData[] = {
                     0, 0,
                     1, 0,
                     1, 1,
